Sort build.ninja contents in relatively stable way.
This makes it significantly easier to diff output across versions,
particularly versions that change the env representation, and thus the
hash.

The current key is (type, path, env?) where type indicates the top-level
output directory (gen, env, latest) and env? is a digest for
context-sensitive products.

Because this is merely a permutation of the contents of build.ninja, it
should have no user-visible impact (no rebuild).