Add concept of target checks.
These are lists of functions (or target-specific methods) run when
a target is evaluated in an environment.  They have access to the
environment after 'using' is applied, which allows them to check
properties of the environment -- for example, to verify that no
transitive dependence on a certain target has been introduced.