Allow env interpolation in deps.
Target dependencies can now be parameterized on the environment, with
some caveats.

Interpolated deps cannot introduce new nodes to the BUILD graph.  They
are evaluated *after* all packages are resolved.  This implies that you
need to ensure that any targets referenced by interpolated deps are also
referenced by at least one static dep (or seed).

Interpolated deps are evaluated in terms of a target's "up-environment"
only, meaning the environment it receives from its dependent and its own
local delta.  Using-deltas in dependencies cannot affect the deps list.

The fact that the C target had to be altered for this change suggests
that deps handling may be factored wrong in general.