Fix fingerprinting to generate reproducible values.
The cPickle module is not guaranteed to generate the same byte stream
for identical objects. Depending on the reference count of an object,
cPickle will store the object in the memo area, and generate a different
string representation (see put opcode). E.g.:

>>> a = ([1],)
>>> cPickle.dumps(a) == cPickle.dumps(([1],))
False

This was causing duplicate products to be generated for a single key.
cPickle fast mode could be enabled which disables the memo area, but it
doesn't decrease cobbling time vs converting the object to a string, and
makes the code slightly more complicated.

This change decreases cobbling time on our project by ~8%.